LayoutTests/svg/dom/SVGException-obsolete.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test.js"></script>
<script>
description("Tests that the obsolete SVGException is not exposed.");

shouldBe("window.SVGException", "undefined");
</script>
</body>
</html>
