Safari-609.1.11.1/LayoutTests/svg/dom/SVGException-obsolete-expected.txt

Tests that the obsolete SVGException is not exposed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.SVGException is undefined
PASS successfullyParsed is true

TEST COMPLETE
